refactor: 날짜포멧팅함수 코어로 분리
